data_collection/requirements.txt
pyserial
flirpy==0.3.0
rplidar-roboticia==0.9.5
openradar
numpy
jupyter
matplotlib
pandas
scikit-learn
scipy
opencv-python-headless==4.6.0.66
requests
python-nmap==0.7.1
soundfile
librosa
streamlit
jsonify
paramiko
# depthai
# openradar
seaborn
imblearn
cleanlab
# xgboost
scikit-image
# hdbscan
keras==2.10.0
mmaction2==0.24.0
soundfile==0.11.0
# tensorflow==2.10.0
# xgboost==1.7.1
jstyleson

# //conda install -c conda-forge python-sounddevice